# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation CodeJobs",
    "description": "Swagger CodeJobs by nel386"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/job/jobs-applied/{loginId}/{jobId}": {
      "delete": {
        "tags": ["Job"],
        "summary": "Borrar una oferta de trabajo a la que se ha aplicado un usuario específico",
        "description": "Endpoint para borrar una oferta de trabajo a la que se ha aplicado un usuario específico",
        "operationId": "removeJobApplication",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "loginId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Oferta de trabajo borrada exitosamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Oferta de trabajo no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/job/employer-jobs/{loginId}": {
      "get": {
        "tags": ["Job"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Obtener todos los trabajos de un empleador por su loginId",
        "description": "Endpoint para obtener todos los trabajos de un empleador por su loginId",
        "operationId": "getEmployerJobsByLoginId",
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "loginId",
            "in": "path",
            "required": true,
            "description": "Id de inicio de sesión del empleador",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de trabajos del empleador"
          },
          "401": {
            "description": "Token no proporcionado o inválido"
          }
        }
      }
    },
...
  }
}